requirements.txt
python-telegram-bot[job-queue]>=20.8,<21.0
notion-client>=2.2.1
APScheduler>=3.10.4,<4.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
